Move fsEndpoint constant
