+++ lib/spaces/constants.ts
@@ -1,2 +1,3 @@
 export const plugPrefix = "_plug/";
 export const stdLibPrefix = "Library/Std/";
+export const fsEndpoint = "/.fs";
