+++ lib/spaces/http_space_primitives.ts
@@ -119,7 +119,7 @@ export class HttpSpacePrimitives implements SpacePrimitives {
   }
 
   async fetchFileList(): Promise<FileMeta[]> {
-    const resp = await this.authenticatedFetch(`${this.url}`, {
+    const resp = await this.authenticatedFetch(this.url, {
       method: "GET",
     });
