+++ web/constants.ts
@@ -1,3 +1,2 @@
 export const maximumDocumentSize = 10; // MiB
 export const defaultLinkStyle = "wikilink";
-export const fsEndpoint = "/.fs";
